Refactor css
