+++ src/Home.js
@@ -2,14 +2,18 @@ import SideNav from "./SideNav";
 import "./styles.css";
 import VideoList from "./VideoList";
 import Nav from "./Nav";
+import NavBar from "./NavBar";
 
 export default function Home() {
   return (
     <div className="App">
       <div className="navbar">
+        <NavBar />
+      </div>
+      <div className="sidenav">
         <SideNav />
       </div>
-      <div className="video">
+      <div className="videolist">
         <Nav />
         <VideoList />
       </div>
