+++ src/PlayVideo.js
@@ -2,6 +2,9 @@ import { useParams } from "react-router-dom";
 import YouTube from "react-youtube";
 import "./css/playvideo.css";
 import SideNav from "./SideNav";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faThumbsUp, faList } from "@fortawesome/free-solid-svg-icons";
+
 export default function PlayVideo() {
   const opts = {
     height: "500px",
@@ -11,11 +14,18 @@ export default function PlayVideo() {
   const { videoId } = useParams();
   return (
     <div className="video-player-container">
-      <div className="sidenav">
-        <SideNav />
-      </div>
+      <SideNav />
+
       <div className="video-player">
         <YouTube videoId={videoId} opts={opts} onReady={this._onReady} />;
+        <div className="util-container">
+          <FontAwesomeIcon className="icon" icon={faThumbsUp} />
+          <p className="icon">
+            {" "}
+            <FontAwesomeIcon icon={faList} />
+            Create playlist
+          </p>
+        </div>
       </div>
     </div>
   );
