+++ src/css/playvideo.css
@@ -1,12 +1,24 @@
 .video-player-container {
   width: 80%;
   height: 100vh;
-  margin: 0 auto;
   display: flex;
-}
-.sidenav {
-  width: 10%;
+  justify-content: end;
+  margin: 0;
 }
 .video-player {
+  display: flex;
   width: 70%;
+  flex-direction: column;
+  justify-content: start;
+}
+.util-container {
+  background-color: #fff;
+  display: flex;
+  width: 100%;
+  padding: 1rem;
+  justify-content: space-between;
+  height: max-content;
+}
+.icon {
+  cursor: pointer;
 }
